+++ packages/spright-components/src/all-components.ts
@@ -12,4 +12,5 @@ import './chat/message';
 import './chat/message/inbound';
 import './chat/message/outbound';
 import './chat/message/system';
+import './icons/all-icons';
 import './rectangle';